third_party/blink/web_tests/http/tests/devtools/elements/edit/undo-set-outer-html-expected.txt

Tests undo for the DOMAgent.setOuterHTML protocol method.


Running: testSetUp

Running: testChangeCharacterData
Replacing 'Getting involved' with 'Getting not involved'

Wrapper identity: identity
Event CharacterDataModified: #text
==========8<==========
<div id="container" style="display:none">
      <p>WebKit is used by <a href="http://www.apple.com/safari/">Safari</a>, Dashboard, etc..</p>
      <h2>Getting not involved</h2>
      <p id="identity">There are many ways to get involved. You can:</p>
      </div>
==========>8==========

Bringing things back

Wrapper identity: identity
Event CharacterDataModified: #text
==========8<==========
<div id="container" style="display:none">
      <p>WebKit is used by <a href="http://www.apple.com/safari/">Safari</a>, Dashboard, etc..</p>
      <h2>Getting involved</h2>
      <p id="identity">There are many ways to get involved. You can:</p>
      </div>
==========>8==========





Running: testChangeAttributes
Replacing '<a href' with '<a foo="bar" href'

Wrapper identity: identity
Event AttrModified: A
Event AttrModified: A
Event AttrRemoved: A
==========8<==========
<div id="container" style="display:none">
      <p>WebKit is used by <a foo="bar" href="http://www.apple.com/safari/">Safari</a>, Dashboard, etc..</p>
      <h2>Getting involved</h2>
      <p id="identity">There are many ways to get involved. You can:</p>
      </div>
==========>8==========

Bringing things back

Wrapper identity: identity
Event AttrModified: A
Event AttrRemoved: A
Event AttrRemoved: A
==========8<==========
<div id="container" style="display:none">
      <p>WebKit is used by <a href="http://www.apple.com/safari/">Safari</a>, Dashboard, etc..</p>
      <h2>Getting involved</h2>
      <p id="identity">There are many ways to get involved. You can:</p>
      </div>
==========>8==========





Running: testRemoveLastChild
Replacing 'Getting involved' with ''

Wrapper identity: identity
Event NodeRemoved: #text
==========8<==========
<div id="container" style="display:none">
      <p>WebKit is used by <a href="http://www.apple.com/safari/">Safari</a>, Dashboard, etc..</p>
      <h2></h2>
      <p id="identity">There are many ways to get involved. You can:</p>
      </div>
==========>8==========

Bringing things back

Wrapper identity: identity
Event NodeInserted: #text
==========8<==========
<div id="container" style="display:none">
      <p>WebKit is used by <a href="http://www.apple.com/safari/">Safari</a>, Dashboard, etc..</p>
      <h2>Getting involved</h2>
      <p id="identity">There are many ways to get involved. You can:</p>
      </div>
==========>8==========





Running: testSplitNode
Replacing 'Getting involved' with 'Getting</h2><h2>involved'

Wrapper identity: identity
Event NodeInserted: H2
Event NodeInserted: H2
Event NodeRemoved: H2
==========8<==========
<div id="container" style="display:none">
      <p>WebKit is used by <a href="http://www.apple.com/safari/">Safari</a>, Dashboard, etc..</p>
      <h2>Getting</h2><h2>involved</h2>
      <p id="identity">There are many ways to get involved. You can:</p>
      </div>
==========>8==========

Bringing things back

Wrapper identity: identity
Event NodeInserted: H2
Event NodeRemoved: H2
Event NodeRemoved: H2
==========8<==========
<div id="container" style="display:none">
      <p>WebKit is used by <a href="http://www.apple.com/safari/">Safari</a>, Dashboard, etc..</p>
      <h2>Getting involved</h2>
      <p id="identity">There are many ways to get involved. You can:</p>
      </div>
==========>8==========





Running: testChangeNodeName
Replacing '<h2>Getting involved</h2>' with '<h3>Getting involved</h3>'

Wrapper identity: identity
Event NodeInserted: H3
Event NodeRemoved: H2
==========8<==========
<div id="container" style="display:none">
      <p>WebKit is used by <a href="http://www.apple.com/safari/">Safari</a>, Dashboard, etc..</p>
      <h3>Getting involved</h3>
      <p id="identity">There are many ways to get involved. You can:</p>
      </div>
==========>8==========

Bringing things back

Wrapper identity: identity
Event NodeInserted: H2
Event NodeRemoved: H3
==========8<==========
<div id="container" style="display:none">
      <p>WebKit is used by <a href="http://www.apple.com/safari/">Safari</a>, Dashboard, etc..</p>
      <h2>Getting involved</h2>
      <p id="identity">There are many ways to get involved. You can:</p>
      </div>
==========>8==========
